baitapthem

+++ module2/src/MVCPerson/untils/ReadFileUntil.java
@@ -24,7 +24,7 @@ public class ReadFileUntil {
         String[] info ;
         for(String line : stringList){
             info =line.split(",");
-            students.add(new Student(Integer.parseInt(info[0]),info[1],Integer.parseInt(info[2])));
+            students.add(new Student(Integer.parseInt(info[0]),info[1],Double.parseDouble(info[2])));
         }
         return students;
      }
